Refactored Splash Screen
