+++ lib/loginscreen.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:geolocator/geolocator.dart';
 import 'package:masjid_pass/settingspage.dart';
+import 'package:masjid_pass/shared_preferences/user_shared_preferences.dart';
+
 import 'db/masjid_database.dart';
 import 'models/user.dart';
-import 'package:geolocator/geolocator.dart';
 
 ///Main fuction to set the preferred orientation of the screen
 Future main() async {
@@ -197,6 +199,7 @@ class _LoginPageState extends State<LoginPage> with TickerProviderStateMixin {
   Future<bool> _loginPressed() async {
     bool? validLogin = await getLogin(username, password);
     if (validLogin == true) {
+      await UserSharedPreferences.setUserLoggedIn(true);
       _navigateToSettings();
       return true;
     }
